scripts/sysmon/requirements.txt
PyYAML==3.12
protobuf==3.0.0
psutil==4.3.1
pyasn1-modules==0.0.8
rsa==3.4.2
